requirements.txt
alembic==1.0.10
amqp==5.0.6
aniso8601==9.0.1
asn1crypto==0.24.0
attrs==21.2.0
autoenv==1.0.0
bcrypt==3.1.7
billiard==3.6.4.0
blinker==1.4
cachetools==3.1.1
celery==5.1.2
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
codecov==2.0.15
coverage==4.5.3
cryptography==2.7
enum34==1.1.6
flasgger==0.9.3
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.20.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-UUID==0.2
google-auth==1.6.3
gunicorn==20.1.0
idna==2.8
iniconfig==1.1.1
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
kombu==5.1.0
kubernetes==10.0.0
Mako==1.0.10
MarkupSafe==1.1.1
marshmallow==2.19.2
mistune==0.8.4
mysql-connector==2.2.9
newrelic==7.0.0.166
nose==1.3.7
oauthlib==3.0.2
packaging==21.3
pluggy==1.0.0
prometheus-http-client==1.0.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyJWT==1.7.1
pyparsing==3.0.7
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.0
python-dotenv==0.10.2
python-editor==1.0.4
pytz==2021.3
pyxdg==0.25
PyYAML==5.1.1
redis==3.5.3
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
SecretStorage==2.3.1
six==1.12.0
SQLAlchemy==1.3.4
SQLAlchemy-Utils==0.36.1
toml==0.10.2
typing==3.6.6
urllib3==1.25.3
vine==5.0.0
virtualenv==16.6.0
wcwidth==0.2.5
websocket-client==0.56.0
Werkzeug==0.15.5
